% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3-methods.R
\name{confint.DTRreg}
\alias{confint.DTRreg}
\title{Confidence Interval Calculations for DTRs}
\usage{
\method{confint}{DTRreg}(object, parm = NULL, level = 0.95, type = c("se", "percentile"), ...)
}
\arguments{
\item{object}{A model object generated by the function DTRreg.}

\item{parm}{Not available for DTRreg objects.}

\item{level}{The confidence level required.}

\item{type}{Typical Wald-type confidence interval "se" (default) or 
confidence intervals derived with the percentile method "percentile" 
(bootstrap variance estimates only).}

\item{...}{Space for additional arguments (not currently used).}
}
\value{
A list with columns giving lower and upper confidence limits for each
  parameter. These will be labelled as (1-level)/2 and 1 - (1-level)/2 in 
  percentage (by default 2.5\% and 97.5\%).
}
\description{
Confidence intervals for parameters, with the option of 
  constructing the confidence intervals using the percentile method 
  when bootstrap is used.
}
